[CI] add continuous integration using AppVeyor
AppVeyor is a CI testing service similar to Travis, but runs
on Windows machines (Windows Server 2012 instances).

We build mexopencv against OpenCV 3.0.0 and Octave 4.0.0,
then run the test suite:
- official Octave binaries used
- custom pre-built OpenCV binaries downloaded (MinGW x86)

mexopencv is compiled using the same GCC compilers. We can use either
the mexopencv.make() script or the makefile to drive the build.

We use octave-cli executable for runnign the tests and ignore the exit code
(due to a bug in the Windows build of Octave)

---

OpenCV (both opencv and opencv_contrib select modules) was built using
MinGW GCC 4.9, the same version bundled with the official Octave 4.0.0
32-bit Windows binaries. This was taken out of the main mexopencv build
process because it takes over 45 min to compile OpenCV alone on AppVeyor
(if we exceed 1 hour the job is killed).