update MxArray <-> cv::SparseMat conversions
Note that on the MATLAB side, it only supports 2D double sparse matrices.
But now we handle any depth of SparseMat on the OpenCV side, both in the
MxArray::MxArray(SparseMat) constructor and the MxArray::toSparse method.

We also update some imgproc functions that deal with sparse histograms
to explicitly specify CV_32F as target depth.